targets: Ultra96 enable hardware preset
Enable the hardware preset specified by the vendor
for both v1 and v2 variants. Both are different and not
compatible with one another due to changes in BOM (DDR).

The user can now specify either "ultra96v1" or "ultra96v2"
variants. The respective board presets will be used. By
default Vivado does not know about the Ultra96 presets.
It is the users responsibility to install the board_files
provided by AVNET prior to building hardware.

https://github.com/Avnet/bdf.git

Download and follow instructions to install board files.

If a preset is not required, simply set the properties
fpga_board = "None" and fpga_board_rev = "None" and a
preset will not be used.